# backend/scratch/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agentic MCP Gateway",
    "description": "AI-powered orchestration layer that connects to multiple third-party services via MCP servers, understands natural language workflow descriptions, decomposes them into executable DAG steps, and orchestrates cross-service operations.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/execute": {
      "post": {
        "tags": [
          "Execution"
        ],
        "summary": "Execute a workflow DAG",
        "description": "Executes a validated DAG through the execution engine. Runs nodes in topological order with parallelism for independent nodes. Returns final execution results with per-node status.",
        "operationId": "execute_workflow_execute_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecuteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecuteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/execute/stream": {
      "post": {
        "tags": [
          "Execution"
        ],
        "summary": "Execute DAG with SSE streaming",
        "description": "Executes a validated DAG and streams real-time events via Server-Sent Events. Events include: workflow_start, node_start, node_running, node_success, node_failed, node_retry, node_skipped, hitl_required, workflow_complete.",
        "operationId": "execute_workflow_stream_execute_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecuteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/execute/status": {
      "get": {
        "tags": [
          "Execution"
        ],
        "summary": "Get execution status",
        "description": "Retrieves the current status and results of a workflow execution by ID.",
        "operationId": "get_execution_status_execute_status_get",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Execution ID",
              "title": "Id"
            },
            "description": "Execution ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Execution Status Execute Status Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/execute/approve/{execution_id}/{node_id}": {
      "post": {
        "tags": [
          "Execution"
        ],
        "summary": "Approve a HITL gate",
        "description": "Approves a pending human-in-the-loop approval for a specific node.",
        "operationId": "approve_hitl_execute_approve__execution_id___node_id__post",
        "parameters": [
          {
            "name": "execution_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Execution Id"
            }
          },
          {
            "name": "node_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Node Id"
            }
          },
          {
            "name": "approved",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Approved"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Approve Hitl Execute Approve  Execution Id   Node Id  Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/status": {
      "get": {
        "tags": [
          "Execution"
        ],
        "summary": "Get Execution Status",
        "description": "Retrieve the current status of a workflow execution by ID.",
        "operationId": "get_execution_status_status_get",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "DAGNode": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Unique node identifier, e.g. 'node_1' or 'task_1'"
          },
          "tool": {
            "type": "string",
            "title": "Tool",
            "description": "Target MCP tool: jira, github, slack, sheets"
          },
          "action": {
            "type": "string",
            "title": "Action",
            "description": "Tool-specific action to invoke"
          },
          "params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params",
            "description": "Action parameters, may contain {{template}} refs"
          },
          "depends_on": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Depends On",
            "description": "IDs of upstream dependency nodes"
          },
          "requires_approval": {
            "type": "boolean",
            "title": "Requires Approval",
            "description": "Whether this node needs HITL approval",
            "default": false
          },
          "retry": {
            "$ref": "#/components/schemas/RetryConfig"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "Human-readable step name"
          },
          "mock_output": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mock Output",
            "description": "Mock output for demo/testing"
          },
          "timeout_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeout Ms",
            "description": "Per-node timeout in milliseconds"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tool",
          "action"
        ],
        "title": "DAGNode",
        "description": "A single step in the workflow DAG."
      },
      "ExecuteRequest": {
        "properties": {
          "dag": {
            "$ref": "#/components/schemas/WorkflowDAG-Input",
            "description": "The DAG to execute"
          },
          "auto_approve": {
            "type": "boolean",
            "title": "Auto Approve",
            "description": "If True, skip HITL approval gates (demo mode)",
            "default": true
          },
          "dry_run": {
            "type": "boolean",
            "title": "Dry Run",
            "description": "If True, simulate execution without calling real tools",
            "default": false
          },
          "credentials": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credentials",
            "description": "User-specific credentials from the ConnectTools dashboard"
          },
          "chat_history": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chat History",
            "description": "Full conversation history to persist"
          }
        },
        "type": "object",
        "required": [
          "dag"
        ],
        "title": "ExecuteRequest",
        "description": "POST /execute \u2014 request body."
      },
      "ExecuteResponse": {
        "properties": {
          "execution_id": {
            "type": "string",
            "title": "Execution Id"
          },
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "total_nodes": {
            "type": "integer",
            "title": "Total Nodes"
          },
          "succeeded": {
            "type": "integer",
            "title": "Succeeded"
          },
          "failed": {
            "type": "integer",
            "title": "Failed"
          },
          "skipped": {
            "type": "integer",
            "title": "Skipped"
          },
          "results": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Results"
          },
          "audit_log": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Audit Log"
          }
        },
        "type": "object",
        "required": [
          "execution_id",
          "success",
          "total_nodes",
          "succeeded",
          "failed",
          "skipped"
        ],
        "title": "ExecuteResponse",
        "description": "POST /execute \u2014 final response (non-streaming)."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RetryConfig": {
        "properties": {
          "max_attempts": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Max Attempts",
            "default": 3
          },
          "backoff_factor": {
            "type": "number",
            "minimum": 1.0,
            "title": "Backoff Factor",
            "default": 2.0
          },
          "initial_delay": {
            "type": "number",
            "minimum": 0.1,
            "title": "Initial Delay",
            "default": 1.0
          },
          "timeout": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Timeout",
            "description": "Per-attempt timeout in seconds",
            "default": 10
          }
        },
        "type": "object",
        "title": "RetryConfig",
        "description": "Per-node retry policy with exponential backoff."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WorkflowDAG-Input": {
        "properties": {
          "workflow_name": {
            "type": "string",
            "title": "Workflow Name",
            "description": "Human-readable workflow name"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/DAGNode"
            },
            "type": "array",
            "minItems": 1,
            "title": "Nodes",
            "description": "Ordered list of DAG steps"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "workflow_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workflow Id",
            "description": "Unique execution ID"
          },
          "execution_layers": {
            "anyOf": [
              {
                "items": {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution Layers",
            "description": "Manually specified execution layers from the planner"
          },
          "context_refs": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Refs",
            "description": "Context template assignments"
          }
        },
        "type": "object",
        "required": [
          "workflow_name",
          "nodes"
        ],
        "title": "WorkflowDAG",
        "description": "Complete workflow directed acyclic graph.\nThis is the primary contract between LLM output and the execution engine."
      },
...
    }
  }
}